add device metadata on sign up
